cryptohome: Delete AuthFactor file securely.
BUG=b:263251251
TEST=tast run localhost:9222 cryptohome.AddRemove*

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4193054
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>

+++ cryptohome/auth_factor/auth_factor_manager.cc
@@ -552,16 +552,23 @@ CryptohomeStatus AuthFactorManager::RemoveAuthFactor(
   }
 
   // Remove the file.
-  if (!platform_->DeleteFile(file_path.value())) {
-    LOG(ERROR) << "Failed to delete from disk auth factor "
-               << auth_factor.label() << " of type "
-               << AuthFactorTypeToString(auth_factor.type()) << " for "
-               << obfuscated_username;
-    return MakeStatus<CryptohomeError>(
-        CRYPTOHOME_ERR_LOC(kLocAuthFactorManagerDeleteFailedInRemove),
-        ErrorActionSet({ErrorAction::kDevCheckUnexpectedState,
-                        ErrorAction::kRetry, ErrorAction::kReboot}),
-        user_data_auth::CRYPTOHOME_ERROR_BACKING_STORE_FAILURE);
+  if (!platform_->DeleteFileSecurely(file_path.value())) {
+    LOG(WARNING) << "Failed to securely delete from disk auth factor "
+                 << auth_factor.label() << " of type "
+                 << AuthFactorTypeToString(auth_factor.type()) << " for "
+                 << obfuscated_username
+                 << ". Attempting to delete without zeroization.";
+    if (!platform_->DeleteFile(file_path.value())) {
+      LOG(ERROR) << "Failed to delete from disk auth factor "
+                 << auth_factor.label() << " of type "
+                 << AuthFactorTypeToString(auth_factor.type()) << " for "
+                 << obfuscated_username;
+      return MakeStatus<CryptohomeError>(
+          CRYPTOHOME_ERR_LOC(kLocAuthFactorManagerDeleteFailedInRemove),
+          ErrorActionSet({ErrorAction::kDevCheckUnexpectedState,
+                          ErrorAction::kRetry, ErrorAction::kReboot}),
+          user_data_auth::CRYPTOHOME_ERROR_BACKING_STORE_FAILURE);
+    }
   }
   LOG(INFO) << "Deleted from disk auth factor label: " << auth_factor.label();
   return OkStatus<CryptohomeError>();
